Add the logic to test if the game is won or drawn
Test if the game is over - for the game to be over, it has to be either
won or drawn.

To check for win conditions - one check for horizontal match, one check
for vertical match, and two checks for both diagonal matches were
placed.

To check for draw condition - one check to test if the multidimensional
array does not contain any element that is an instance of an integer
needs to be placed.